Implemented protection route, resorted files
