+++ frontend/src/components/BrandedMembers/style/BrandedMembers.module.css
@@ -50,4 +50,4 @@
     display: none;
   }
 }
-/*# sourceMappingURL=MembersTable.module.css.map */
+/*# sourceMappingURL=BrandedMembers.module.css.map */
